+++ src/no/ntnu/fp/model/record/ActiveRom.java
@@ -18,29 +18,25 @@ import no.ntnu.fp.model.Rom;
 
 import org.apache.derby.tools.sysinfo;
 
-public class ActiveRom {
+public class ActiveRom extends ActiveModel{
 	private static String db_url = "Kal";
 	private static String admin_name = "";
 	private static String admin_pwd = "";
 	protected static Connection connection ;
 	
 	public static void createRom(Rom rom){
-		PreparedStatement ps = null;
-		
 		try{
 			connect();
-			
-			ps = connection.prepareStatement(
-				"INSERT INTO Rom(romID, navn)" +
-				"VALUES ( ?, ? )" 
-			);
-			ps.setInt(1, rom.getRomId());
-			ps.setString(2, rom.getNavn());
-						
-			boolean det_gikk_bra = ps.execute();
-			if (det_gikk_bra){
-				System.out.println("Det gikk bra!");
+			if(connection != null){
+				PreparedStatement ps = connection.prepareStatement(
+					"INSERT INTO Rom(romID, navn)" +
+					"VALUES ( ?, ? )" 
+				);
+				ps.setInt(1, rom.getRomId());
+				ps.setString(2, rom.getNavn());
+				ps.execute();
 			}
+			connection.close();
 		}
 		catch(SQLException e){
 			System.out.println("Kan ikke lage person. Feilmelding:" + e.getMessage());
@@ -65,6 +61,7 @@ public class ActiveRom {
 					romId = rs.getInt("romID");
 				}
 			}
+			connection.close();
 		}
 	    catch( SQLException e){
 	    	System.out.println("Kan ikke finne rom med id = " + romId);
@@ -75,31 +72,24 @@ public class ActiveRom {
 	
 
 	public static void updateRom(Rom rom){
-		PreparedStatement ps = null;
-		
 		String navn = rom.getNavn();
 		int romID = rom.getRomId();
 		
 		try {
         	connect();
         	if( connection != null){
-	            ps = connection.prepareStatement(
-	            		"UPDATE Rom " + 
-	            		"SET navn= ? " +
-	                    "WHERE romID = ? "
+	            PreparedStatement ps = connection.prepareStatement(
+	            		"UPDATE Rom SET navn= ? WHERE romID = ? "
 	            );
 	            ps.setString(1, navn);
 	            ps.setInt(4, romID);
+	            ps.executeUpdate();
         	}
-        	
-            ps.executeUpdate();    
+        	connection.close();
         }
 		catch (SQLException e){
         	System.out.println("Kan ikke oppdatere!");
         	System.out.println("Detaljer:" + e.getMessage());
-        }
-        finally {
-            //if (ps != null) ps.close();
         }
 	}
 	
@@ -111,16 +101,17 @@ public class ActiveRom {
 		
 		try{
 			connect();
-			
-			PreparedStatement ps = connection.prepareStatement(
-					"SELECT * FROM Rom WHERE romID = ? "
-			);
-			ps.setInt(1, romID);
-			
-			ResultSet rs = ps.executeQuery(); 
-			if (rs != null){
-				while(rs.next()){
-					navn = rs.getString("navn");
+			if( connection != null){
+				PreparedStatement ps = connection.prepareStatement(
+						"SELECT * FROM Rom WHERE romID = ? "
+				);
+				ps.setInt(1, romID);
+				
+				ResultSet rs = ps.executeQuery(); 
+				if (rs != null){
+					while(rs.next()){
+						navn = rs.getString("navn");
+					}
 				}
 			}
 		}
@@ -138,27 +129,17 @@ public class ActiveRom {
 	private static void deleteRom(int romID) {
 		try {
 			connect();
-			PreparedStatement ps = connection.prepareStatement(
-					"DELETE FROM Rom WHERE romID = ?"
-			);
-			ps.setInt(1, romID);
-			ps.execute();	
+			if( connection != null){
+				PreparedStatement ps = connection.prepareStatement(
+						"DELETE FROM Rom WHERE romID = ?"
+				);
+				ps.setInt(1, romID);
+				ps.execute();
+			}	
 		} 
 		catch (SQLException e) {
 			e.printStackTrace();
 		}
-	}
-	
-	private static void connect() throws SQLException{
-		try{
-			connection =  DriverManager.getConnection("jdbc:derby:" + db_url, admin_name, admin_pwd );
-		}
-		catch(SQLException e){
-			connection = null;
-			System.out.println("Kan ikke koble til database");
-		}
-		finally{ 
-		}		
 	}
 	
 	public static void main(String args[]){
@@ -174,8 +155,21 @@ public class ActiveRom {
 	
 	
 	
-
-
+	/******************************
+	*  			Tester			  *
+	******************************/
+	
+	private static void testCreateRom(){
+		Rom rom = new Rom("");
+		rom.setRomId(113);
+		rom.setNavn("Martin");
+		
+		createRom(rom);
+		
+//		Rom nyeRom = selectRom(rom.getRomId());
+//		System.out.println(nyeRom.getNavn());
+	}
+	
 	private void testUpdateRom(){
 		int ansattnr = 10001;
 		String navn = "Martin";
@@ -194,22 +188,9 @@ public class ActiveRom {
 		System.out.println("Nytt navn:" + oppdatertPerson.getNavn());
 	}
 	
-	
-	private static void testCreateRom(){
-		Rom rom = new Rom("");
-		rom.setRomId(113);
-		rom.setNavn("Martin");
-		
-		createRom(rom);
-		
-//		Rom nyeRom = selectRom(rom.getRomId());
-//		System.out.println(nyeRom.getNavn());
-	}
-	
-	
-	private void testSelectPerson(){
-		int romID = 113;
-		Rom testRom  = selectRom(romID);
-		System.out.println(testRom.getNavn());
-	}
+//	private void testSelectPerson(){
+//		int romID = 113;
+//		Rom testRom  = selectRom(romID);
+//		System.out.println(testRom.getNavn());
+//	}
 }
